Invalidate mangled name of TType if the type is changed
Sometimes, the mangled name of a type is already generated when the type
is still being changed. For example, this could happen when the type of
A[i] is generated by copying the type of A (already mangled) and calling
clearArrayness() on the copy. Make sure that the mangled name gets
updated if the type is changed by clearing the mangled name when
something affecting it is changed.

BUG=angleproject:1170, 538692
TEST=angle_unittests

Change-Id: I9dfacea653f56536e1573a6dbf60ff21da7cc5ab